requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
braceexpand==0.1.7
Brotli==1.0.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
discord-webhook==0.14.0
discum==1.1.0
dnspython==1.16.0
easygui==0.98.2
emoji==1.6.1
fade==0.0.9
filetype==1.0.8
gevent==21.8.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httplib2==0.20.2
httpx==0.20.0
httpx-socks==0.5.0
idna==3.2
import-expression==1.1.4
jishaku==2.3.0
multidict==5.2.0
outcome==1.1.0
Pillow==8.4.0
protobuf==3.19.1
psutil==5.8.0
pyarmor==7.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.12.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pypiwin32==223
pypresence==4.2.0
PySocks==1.7.1
python-socks==1.2.4
pywin32==302
random-user-agent==1.0.1
requests==2.26.0
requests-futures==1.0.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
spintax==1.0.4
tasksio==0.0.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.7
websocket==0.2.1
websocket-client==1.2.1
win10toast==0.9
wsproto==1.0.0
yarl==1.7.0
youtube-dl==2021.6.6
zope.event==4.5.0
zope.interface==5.4.0